attached_assets/requirements_1760594113262.txt
instagrapi==1.17.9
google-generativeai==0.7.2
